python/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
limits==5.5.0
mysql-connector-python==9.4.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.20
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wrapt==1.17.3
